채팅 관련 db, 코드 수정

+++ src/main/java/com/TingTing/dto/ConditionResponseDto.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class ConditionResponseDto {
-    private Long conditionId;
+    private int conditionId;
 }
